NST-38 Consolidate files
* Update logging library
