+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "@jspsych/plugin-instructions": "^1.1.4",
     "@jspsych/plugin-survey-html-form": "^2.1.0",
+    "consola": "^3.4.2",
     "jspsych": "^7.3.4",
     "neurocog": "^1.1.1"
   },
